Added capability check wrapper methods.
These may seem like a weird addition, but there are a few reasons for
them. The first is that it helps to make the code more readable. While
any dev worth their salt can understand these checks, the method is much
more readable than the actual check. Additionally, this can help future
proof the API. If Forge were to change the way you check for a
capability, I could fix that on my end, and the mods using Tesla would
never know there was an issue. This is the same idea behind private
field design patterns.